move component tests
